\name{central.quantile}
\alias{central.quantile}
\title{Central quantile}
\usage{
central.quantile(x, pct = 0.95)
}
\arguments{
  \item{x}{numeric vector}

  \item{pct}{central percentile desired}
}
\value{
a two element vector giving the lower and upper quantiles
}
\description{
Upper and lower values of central quantile
}
\author{
Eric Archer <eric.archer@noaa.gov>
}

