\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric Mean Function}
\usage{
geometric.mean(x)
}
\arguments{
  \item{x}{vector}
}
\value{
a number giving the geometric mean of a vector
}
\description{
Calculates the geometric mean of a vector
}
\author{
Eric Archer <eric.archer@noaa.gov>
}

