\name{harmonic.mean}
\alias{harmonic.mean}
\title{Harmonic Mean Function}
\usage{
harmonic.mean(x)
}
\arguments{
  \item{x}{vector}
}
\value{
a number giving the harmonic mean of a vector
}
\description{
Calculates the harmonic mean of a vector
}
\examples{
x <- rlnorm(100)
mean(x)
median(x)
harmonic.mean(x)
}
\author{
Eric Archer <eric.archer@noaa.gov>
}

