\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector}
\usage{
normalize(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\value{
a numeric vector of the same length as x
}
\description{
Normalize a numeric vector to have a mean of zero and a
standard deviation of one
}
\author{
Eric Archer <eric.archer@noaa.gov>
}

