\name{copy.tri}
\alias{copy.tri}
\title{Copy matrix triangles}
\usage{
copy.tri(mat, from = "lower", to = "upper")
}
\arguments{
  \item{mat}{a matrix}

  \item{from,to}{triangle to copy from and to. Can be
  "lower" or "upper"}
}
\value{
a matrix
}
\description{
Copy between lower left and upper right triangles of a
matrix
}
\author{
Eric Archer <eric.archer@noaa.gov>
}

